Sekiseigumi Barracks aetheryte fix
When self.initiatedPos is an empty table, IsNull returns an empty table, because {}==nil is false.  This causes a nil arithmetic error when math.distance3d receives this empty table as its first param.

Instead, use Player.pos when self.initiatedPos is empty according to the minionlib table.valid function.