Add ability to display entity id in Scanner.
Separated contentid from name in scanner and made it a separate column.

Added checkboxes to control display of entity id or contentid in Scanner.

Added ability to select rows of Scanner and Rawdata so that when they are clicked the data comprising the row is output to the Minion Console.  This allows the row data to be copy and pasted.